Improve test for Magic Folder statistics and move it from test_system.py to test_magic_folder.py.
